Migrate image picking to separate functions
