+++ app/src/main/java/dev/sanskar/nero/ui/add/AddScreen.kt
@@ -1,10 +1,6 @@
 package dev.sanskar.nero.ui.add
 
-import android.net.Uri
 import androidx.activity.compose.BackHandler
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedContentScope
 import androidx.compose.animation.AnimatedVisibility
@@ -12,12 +8,7 @@ import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.with
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -25,20 +16,11 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.Button
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.OutlinedTextField
-import androidx.compose.material.Surface
 import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.BrowseGallery
-import androidx.compose.material.icons.filled.Camera
-import androidx.compose.material.icons.filled.PhotoAlbum
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.derivedStateOf
@@ -46,28 +28,23 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Alignment.Companion.CenterHorizontally
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.scale
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
 import androidx.hilt.navigation.compose.hiltViewModel
 import coil.compose.AsyncImage
 import dev.sanskar.nero.R
 import dev.sanskar.nero.data.Book
 import dev.sanskar.nero.ui.components.BookRow
+import dev.sanskar.nero.util.PickImage
 import dev.sanskar.nero.util.clickWithRipple
-import dev.sanskar.nero.util.getNewFileUri
-import timber.log.Timber
 
 @OptIn(ExperimentalAnimationApi::class)
 @Composable
@@ -76,9 +53,8 @@ fun AddScreen(
     viewModel: AddViewModel = hiltViewModel(),
     onSelected: () -> Unit
 ) {
-    var isAddingCustomBook by remember { mutableStateOf(false) }
     AnimatedContent(
-        targetState = isAddingCustomBook,
+        targetState = viewModel.isAddingCustomBook,
         transitionSpec = {
             slideIntoContainer(AnimatedContentScope.SlideDirection.Up) +
                     fadeIn() with slideOutOfContainer(AnimatedContentScope.SlideDirection.Down) + fadeOut()
@@ -88,11 +64,11 @@ fun AddScreen(
         if (it) {
             AddCustomBookContent(viewModel, onSelected = onSelected)
             BackHandler {
-                isAddingCustomBook = false
+                viewModel.isAddingCustomBook = false
             }
         } else {
             SearchContent(viewModel, onSelected = onSelected) {
-                isAddingCustomBook = true
+                viewModel.isAddingCustomBook = true
             }
         }
     }
@@ -108,68 +84,14 @@ fun AddCustomBookContent(
     var authorName by remember { mutableStateOf("") }
     var totalPageCount by remember { mutableStateOf("") }
     var categories by remember { mutableStateOf("") }
-
-    val context = LocalContext.current
-
     var imageUri by remember { mutableStateOf("") }
-    var cameraUri by remember { mutableStateOf<Uri?>(null) }
+
     var showImagePicker by remember { mutableStateOf(false) }
 
-    val galleryPicker = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.PickVisualMedia()
-    ) {
-        imageUri = it?.toString() ?: ""
-        Timber.d("Received URI: $it")
-    }
-
-
-    val cameraPicker = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.TakePicture()
-    ) {
-        imageUri = if (it) {
-            cameraUri?.toString() ?: ""
-        } else {
-            ""
-        }
-    }
-
-    if (showImagePicker) Dialog(
-        onDismissRequest = { showImagePicker = false },
-    ) {
-        Surface(
-            modifier = Modifier.fillMaxWidth(0.8f)
-                .height(72.dp),
-            shape = RoundedCornerShape(4.dp)
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.SpaceEvenly
-            ) {
-                IconButton(onClick = {
-                    showImagePicker = false
-                    cameraUri = context.getNewFileUri()
-                    cameraPicker.launch(cameraUri)
-                }) {
-                    Icon(
-                        imageVector = Icons.Default.Camera,
-                        contentDescription = "Pick Image using Camera",
-                        modifier = Modifier
-                            .scale(1.5f)
-                    )
-                }
-
-                IconButton(onClick = {
-                    showImagePicker = false
-                    galleryPicker.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
-                }) {
-                    Icon(
-                        imageVector = Icons.Default.PhotoAlbum,
-                        contentDescription = "Pick Image using Gallery",
-                        modifier = Modifier
-                            .scale(1.5f)
-                    )
-                }
-            }
+    if (showImagePicker) {
+        PickImage {
+            showImagePicker = false
+            imageUri = it
         }
     }
